// src/app/core/models/SanadRahniGheyrManghol/SanadRahniGheyrManghol.ts
export interface SanadRahniGheyrManghol {
    Id?: number;
    ShomarehSanadRahni?: string;
    HozehSabtiDaftarkhanehTanzimkonandehId?: number;
    ShomarehDaftarkhanehTanzimkonandeh?: number;
    SanadRahniDate?: string;
    SanadRahniGheyrMangholTypeId?: number;
    
    MizanArzyabiAvaliyeh?: number;
    MizanAkharinArzyabi?: number;
    MablaghSanadRahni?: number;

    CityName?: string;
    CityId?: number
    ProvinceName?: string;
    ProvinceId?: any;
    ProvinceCode?: any
}